Split up Spell and Utility service methods from item creation
